create and delete with azure storage image for rooms.

+++ HostMaster/HostMaster.Frontend/Pages/Rooms/RoomsIndex.razor.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Components;
 using Microsoft.Extensions.Localization;
 using MudBlazor;
 using HostMaster.Frontend.Shared;
+using HostMaster.Shared.DTOs;
 
 namespace HostMaster.Frontend.Pages.Rooms;
 
@@ -16,12 +17,18 @@ public partial class RoomsIndex
 {
     private List<Room>? Rooms { get; set; }
     private MudTable<Room> table = new();
-    private readonly int[] pageSizeOptions = { 1, 10, 25, 50, int.MaxValue };
+    private readonly int[] pageSizeOptions = { 10, 25, 50, int.MaxValue };
     private int totalRecords = 0;
     private bool loading;
     private const string baseUrl = "api/rooms";
     private string infoFormat = "{first_item}-{last_item} => {all_items}";
     private int accommodationId = 1;
+
+    private bool arrows = true;
+    private bool bullets = true;
+    private bool enableSwipeGesture = true;
+    private bool autocycle = true;
+    private Transition transition = Transition.Slide;
 
     [Inject] private IStringLocalizer<Literals> Localizer { get; set; } = null!;
     [Inject] private IRepository Repository { get; set; } = null!;
@@ -133,6 +140,16 @@ public partial class RoomsIndex
             return;
         }
 
+        var responseHttp3 = await Repository.DeleteAsync($"api/roomphotos/by-roomId/{room.Id}");
+        if (responseHttp3.Error)
+        {
+            var mensajeError = await responseHttp3.GetErrorMessageAsync();
+            Snackbar.Add(Localizer[mensajeError!], Severity.Error);
+            return;
+        }
+
+
+
         var responseHttp = await Repository.DeleteAsync($"{baseUrl}/{room.Id}");
         if (responseHttp.Error)
         {
@@ -147,6 +164,9 @@ public partial class RoomsIndex
             }
             return;
         }
+
+        
+
         await LoadTotalRecordsAsync();
         await table.ReloadServerData();
         Snackbar.Add(Localizer["RecordDeletedOk"], Severity.Success);
